refactor: code clean up and add doc comments
