+++ src/Common/SegmentedControl/types.ts
@@ -1,44 +1,79 @@
-/*
- * Copyright (c) 2024. Devtron Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { OptionType } from '@Common/Types'
+import { TooltipProps } from '@Common/Tooltip'
+import { IconsProps, SelectPickerOptionType } from '@Shared/Components'
 import { ComponentSizeType } from '@Shared/constants'
-import { ReactNode } from 'react'
+import { RefObject } from 'react'
 
-export enum SegmentedControlVariant {
-    // NOTE: values are css class names
-    GRAY_ON_WHITE = 'gui-yaml-switch',
-    WHITE_ON_GRAY = 'gui-yaml-switch-window-bg',
-}
+type SegmentTooltipProps = Omit<
+    TooltipProps,
+    'alwaysShowTippyOnHover' | 'showOnTruncate' | 'shortcutKeyCombo' | 'placement'
+>
 
-export interface SegmentedControlProps {
-    tabs: OptionType<string, ReactNode>[]
-    initialTab: string
-    onChange: (e: React.ChangeEvent<HTMLInputElement>) => void
+export type SegmentType = Pick<SelectPickerOptionType, 'value'> & {
+    /**
+     * If true, the segment will be in error state with error icon
+     */
+    isError?: boolean
+} & (
+        | ({
+              /**
+               * Label for the segment
+               *
+               * Note: Either of label or icon is required
+               */
+              label: SelectPickerOptionType['label']
+              /**
+               * Icon for the segment
+               *
+               * Note: Either of label or icon is required
+               */
+              icon?: IconsProps['name']
+              /**
+               * Tooltip props for the segment
+               *
+               * Note: Required if only icon is provided
+               */
+              tooltipProps?: SegmentTooltipProps
+              ariaLabel?: never
+          } & Pick<SelectPickerOptionType, 'label'>)
+        | {
+              label?: never
+              tooltipProps: SegmentTooltipProps
+              icon: IconsProps['name']
+              /**
+               * Aria label for the segment
+               */
+              ariaLabel: string
+          }
+    )
+
+export type SegmentedControlProps = {
+    /**
+     * List of segments to be displayed
+     */
+    segments: SegmentType[]
     name: string
-    tooltips?: string[]
-    disabled?: boolean
-    rootClassName?: string
-    variant?: SegmentedControlVariant
-    /**
-     * @default ComponentSizeType.medium
-     */
-    size?: ComponentSizeType.medium | ComponentSizeType.large
-    /**
-     * @default false
-     */
-    isControlled?: boolean
+    size?: Extract<ComponentSizeType, ComponentSizeType.xs | ComponentSizeType.small | ComponentSizeType.medium>
+    fullWidth?: boolean
+} & (
+    | {
+          value?: never
+          /**
+           * On change handler for the component
+           */
+          onChange?: (selectedSegment: SegmentType) => void
+      }
+    | {
+          /**
+           * If defined, the component is controlled and onChange needs to be handled by the parent
+           */
+          value: SegmentType['value']
+          onChange: (selectedSegment: SegmentType) => void
+      }
+)
+
+export interface SegmentProps
+    extends Required<Pick<SegmentedControlProps, 'name' | 'onChange' | 'fullWidth' | 'size'>> {
+    isSelected: boolean
+    segment: SegmentType
+    selectedSegmentRef: RefObject<HTMLDivElement> | undefined
 }
